# Allow conditional aggregation