--- a/test/collect_tests.jl
+++ b/test/collect_tests.jl
@@ -46,7 +46,8 @@
         @test aggname(:weight, mean) == "mean_weight"
         @test aggname(x_position, length) == "length_x_position"
         @test aggname((x_position, length)) == "length_x_position"
-        @test aggname((x_position, length, a->a.pos[1]>5)) == "conditional_length_x_position"
+        ypos(a) = a.pos[2]>5
+        @test aggname((x_position, length, ypos)) == "length_x_position_ypos"
     end
 
     @testset "Aggregate Collections" begin
@@ -82,29 +83,31 @@
         @test propertynames(df) == [:step, :mean_weight, :mean_x_position]
         @test df[1, aggname(x_position, mean)] ≈ 4.3333333
 
-        props = [(:weight, mean, a->x_position(a) >= 5)]
-        df = init_agent_dataframe(model, props)
-        collect_agent_data!(df, model, props, 1)
-        @test size(df) == (1, 2)
-        @test propertynames(df) == [:step, :conditional_mean_weight]
-        @test_throws ArgumentError df[1, aggname(:weight, mean)]
-        @test df[1, aggname(:weight, mean; condition = true)] ≈ 0.35
-
+        xtest(agent) = agent.pos[1] > 5
         ytest(agent) = agent.pos[2] > 5
+
         props = [(:weight, mean, ytest)]
         df = init_agent_dataframe(model, props)
         collect_agent_data!(df, model, props, 1)
         @test size(df) == (1, 2)
-        @test propertynames(df) == [:step, :conditional_mean_weight]
+        @test propertynames(df) == [:step, :mean_weight_ytest]
         @test df[1, aggname((:weight, mean, ytest))] ≈ 0.67
 
         props = [(:weight, mean), (:weight, mean, ytest)]
         df = init_agent_dataframe(model, props)
         collect_agent_data!(df, model, props, 1)
         @test size(df) == (1, 3)
-        @test propertynames(df) == [:step, :mean_weight, :conditional_mean_weight]
+        @test propertynames(df) == [:step, :mean_weight, :mean_weight_ytest]
         @test df[1, aggname(:weight, mean)] ≈ 0.37333333333
-        @test df[1, aggname(:weight, mean; condition = true)] ≈ 0.67
+        @test df[1, aggname(:weight, mean, ytest)] ≈ 0.67
+
+        props = [(:weight, mean, xtest), (:weight, mean, ytest)]
+        df = init_agent_dataframe(model, props)
+        collect_agent_data!(df, model, props, 1)
+        @test size(df) == (1, 3)
+        @test propertynames(df) == [:step, :mean_weight_xtest, :mean_weight_ytest]
+        @test df[1, aggname(:weight, mean, xtest)] ≈ 0.35
+        @test df[1, aggname(:weight, mean, ytest)] ≈ 0.67
     end
 
     @testset "High-level API for Collections" begin