Closes #275.

The only issue I see here is that we don't separate our naming convention. There's no indication that `mean_weight` is over all or just some agents. Should we extend aggname and our column naming function to include something like `:conditional_mean_weight`? 

## Benchmarks

Before changes:
```julia
julia> props = [(:weight, mean)]
1-element Array{Tuple{Symbol,typeof(mean)},1}:
 (:weight, Statistics.mean)

julia> @benchmark collect_agent_data!(df, model, props, 1)
BenchmarkTools.Trial: 
  memory estimate:  128 bytes
  allocs estimate:  8
  --------------
  minimum time:     220.822 ns (0.00% GC)
  median time:      247.047 ns (0.00% GC)
  mean time:        332.252 ns (18.49% GC)
  maximum time:     192.535 μs (99.86% GC)
  --------------
  samples:          10000
  evals/sample:     506
```

After:
```julia
julia> props = [(:weight, mean)]
1-element Array{Tuple{Symbol,typeof(mean)},1}:
 (:weight, Statistics.mean)

julia> @benchmark collect_agent_data!(df, model, props, 1)
BenchmarkTools.Trial: 
  memory estimate:  144 bytes
  allocs estimate:  9
  --------------
  minimum time:     214.961 ns (0.00% GC)
  median time:      229.436 ns (0.00% GC)
  mean time:        296.531 ns (12.78% GC)
  maximum time:     122.061 μs (99.77% GC)
  --------------
  samples:          10000
  evals/sample:     545
```

With conditional aggregation:
```julia
julia> ytest(agent) = agent.pos[2] > 5
ytest (generic function with 1 method)

julia> props = [(:weight, mean, ytest)]
1-element Array{Tuple{Symbol,typeof(mean),typeof(ytest)},1}:
 (:weight, Statistics.mean, ytest)

julia> @benchmark collect_agent_data!(df, model, props, 1)
BenchmarkTools.Trial: 
  memory estimate:  144 bytes
  allocs estimate:  6
  --------------
  minimum time:     151.653 ns (0.00% GC)
  median time:      167.491 ns (0.00% GC)
  mean time:        240.460 ns (18.29% GC)
  maximum time:     107.437 μs (99.81% GC)
  --------------
  samples:          10000
  evals/sample:     825
```