@Libbum 2020-08-14
Just added a failing test to show a (contrived) example of what may happen if we don't do something about aggname.