@Datseris 2020-08-14
My only beef with the current naming scheme `conditional_...` is that it doesn't allow 
```julia
ytest(agent) = agent.pos[2] > 5
ytest2(agent) = agent.pos[2] <= 5
props = [(:weight, mean, ytest), (:weight, mean, ytest2)]
```
Why not continue doing what we do already, and just add a third string to `props`, so `join(string(a) for a in tuple, '_')`