@Libbum 2020-08-14
Sure, we can do that. I guess we already have the warning about anonymous functions in the docs. I'll drop the anonymous tests and change this.