@Libbum 2020-08-14
What about a naming convention like this: "mean_weight|x_pos", where you have agg_k|cond, so that you can know that this average is 'dependent on `x_pos`'?