@Datseris 2020-08-14
It seems like a nice name, but I still have in my head the convseration we had some time ago with a user that convinced me to change the convention from `agg(k)` to `agg_k`... but I don't remember in which issue it was! Maybe something similar would happen in the current case of using `|`? I need to find this conversation, because I don't remember what was the problem.