@Libbum 2020-08-14
Ah. Yes, we will have to do `mean_weight_x_pos`, since you cannot have some dataframe `result` and call `result.mean_weight|x_pos`.