@Datseris 2020-08-14
> `result.mean_weight|x_pos`

But this is not valid syntax with respect to the dataframes API, right? Because I think their syntax now is `result[!, "mean_weight|x_pos"]`. In fact it could be that they moved into strings from symbols for such weird column names.